# frontend/src/types/schemas/app-openapi.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"string"}}}}},"tags":["App"]}},"/servers":{"post":{"description":"Creates a new Server and returns its public view","operationId":"ServerController_createServer","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateServerDto"}}}},"responses":{"201":{"description":"The created Server","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerViewDto"}}}}},"summary":"Create a new Server","tags":["servers"]},"get":{"operationId":"ServerController_listServer","parameters":[{"name":"q","required":false,"in":"query","description":"free Text search","schema":{"type":"string"}},{"name":"degreeSlug","required":false,"in":"query","description":"Degree slug","schema":{"type":"string"}},{"name":"degreeId","required":false,"in":"query","description":"Degree identifier","schema":{"type":"string"}},{"name":"startYear","required":false,"in":"query","description":"Student start year (calendar year)","schema":{"minimum":1900,"maximum":2100,"type":"number"}},{"name":"paginated","required":false,"in":"query","description":"Boolean value to indicate if the query is paginated or not","schema":{"type":"boolean"}},{"name":"page","required":false,"in":"query","schema":{"minimum":1,"default":1,"type":"number"}},{"name":"pageSize","required":false,"in":"query","schema":{"minimum":1,"maximum":100,"default":20,"type":"number"}},{"name":"type","required":false,"in":"query","description":"Select Type ","schema":{"$ref":"#/components/schemas/ServerType"}}],"responses":{"200":{"description":"Paginated list of servers","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerListResponseDto"}}}}},"summary":"lists all the servers in a particular type","tags":["servers"]}},"/servers/{serverId}":{"get":{"operationId":"ServerController_getServer","parameters":[{"name":"serverId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"The requested Server","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerViewDto"}}}}},"summary":"Retrieve a server by id","tags":["servers"]},"patch":{"description":"This updates the existing Server and return its public view","operationId":"ServerController_update","parameters":[{"name":"serverId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateServerDto"}}}},"responses":{"200":{"description":"The updated Server","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ServerViewDto"}}}}},"summary":"updating an existing Server","tags":["servers"]}},"/servers/{serverId}/owners":{"post":{"operationId":"ServerController_addOwner","parameters":[{"name":"serverId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddOwnerDto"}}}},"responses":{"200":{"description":"Owner added successfully"},"201":{"description":""}},"summary":"Add an owner to a server by email","tags":["servers"]}},"/servers/{serverId}/owners/{userId}":{"delete":{"operationId":"ServerController_removeOwner","parameters":[{"name":"serverId","required":true,"in":"path","schema":{"type":"string"}},{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Owner removed successfully"}},"summary":"Remove an owner from a server","tags":["servers"]}},"/servers/{serverId}/me":{"get":{"operationId":"ServerController_myRoles","parameters":[{"name":"serverId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Returns the roles array for the current user in this server","content":{"application/json":{"schema":{"type":"object","properties":{"roles":{"type":"array","items":{"type":"string"}}},"required":["roles"]}}}}},"summary":"Get the current user membership roles for a server","tags":["servers"]}},"/notifications":{"post":{"operationId":"NotificationContoller_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateNotificationDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationViewDto"}}}}},"summary":"Creates notificaion in the database for the specified user","tags":["NotificationContoller"]}},"/notifications/unread":{"get":{"operationId":"NotificationContoller_listUnread","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/NotificationViewDto"}}}}}},"summary":"List unread notifications for the current user","tags":["NotificationContoller"]}},"/notifications/{notificationId}/read":{"patch":{"operationId":"NotificationContoller_markRead","parameters":[{"name":"notificationId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationViewDto"}}}}},"summary":"Mark a notification as read","tags":["NotificationContoller"]}},"/notifications/preferences/channels/{channelId}":{"get":{"operationId":"NotificationContoller_getChannelPreference","parameters":[{"name":"channelId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationPreferenceViewDto"}}}}},"summary":"Get the current user's notification preference for a channel","tags":["NotificationContoller"]},"put":{"operationId":"NotificationContoller_setChannelPreference","parameters":[{"name":"channelId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateNotificationPreferenceDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/NotificationPreferenceViewDto"}}}}},"summary":"Update the current user's notification preference for a channel","tags":["NotificationContoller"]}},"/notifications/stream":{"get":{"operationId":"NotificationContoller_stream","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"object"}}}}},"tags":["NotificationContoller"]}},"/servers/{serverId}/channels":{"post":{"operationId":"ChannelsController_create","parameters":[{"name":"serverId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateChannelDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChannelViewDto"}}}}},"tags":["Channels"]},"get":{"operationId":"ChannelsController_list","parameters":[{"name":"serverId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"returns a channel list visible to the user","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChannelListResponseDto"}}}}},"tags":["Channels"]}},"/servers/{serverId}/channels/{channelId}":{"get":{"operationId":"ChannelsController_find","parameters":[{"name":"channelId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"returns the desired channel","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChannelViewDto"}}}}},"tags":["Channels"]},"delete":{"operationId":"ChannelsController_deleteChannel","parameters":[{"name":"channelId","required":true,"in":"path","schema":{"type":"string"}},{"name":"serverId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"Deletes a channel and all its access records"}},"tags":["Channels"]}},"/servers/{serverId}/channels/{channelId}/members":{"post":{"operationId":"ChannelsController_addMembers","parameters":[{"name":"channelId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"201":{"description":"returns when adding a member in ther channel if the channel is hidden"}},"tags":["Channels"]}},"/servers/{serverId}/channels/{channelId}/members/{userId}":{"delete":{"operationId":"ChannelsController_removeMember","parameters":[{"name":"channelId","required":true,"in":"path","schema":{"type":"string"}},{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"this handler deletes a member from the channel"}},"tags":["Channels"]}},"/servers/{serverId}/channels/{channelId}/messages":{"post":{"operationId":"MessagesController_create","parameters":[{"name":"serverId","required":true,"in":"path","schema":{"type":"string"}},{"name":"channelId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMessageDto"}}}},"responses":{"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageViewDto"}}}}},"summary":"Create a new message in the channel","tags":["messages"]},"get":{"operationId":"MessagesController_listMessages","parameters":[{"name":"serverId","required":true,"in":"path","schema":{"type":"string"}},{"name":"channelId","required":true,"in":"path","schema":{"type":"string"}},{"name":"page","required":true,"in":"query","schema":{"type":"string"}},{"name":"pageSize","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageListResponseDto"}}}}},"summary":"List messages in the channel","tags":["messages"]}},"/degrees":{"get":{"operationId":"DegreeController_listDegrees","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DegreeViewDto"}}}}}},"summary":"List all degrees","tags":["degrees"]}},"/degrees/{slug}":{"get":{"operationId":"DegreeController_getDegree","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DegreeDetailDto"}}}}},"summary":"Retrieve degree details by slug","tags":["degrees"]}},"/degrees/{slug}/modules":{"get":{"operationId":"DegreeController_listModules","parameters":[{"name":"slug","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/DegreeModuleViewDto"}}}}}},"summary":"List modules for a degree","tags":["degrees"]}}},"info":{"title":"Campus Hub API","description":"REST API for Campus Hub","version":"1.0.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"ServerType":{"type":"string","enum":["unimodules","citysocieties","personal"]},"CreateServerDto":{"type":"object","properties":{"name":{"type":"string","minLength":2,"maxLength":60,"example":"Calisthenics","description":"We discuss calisthenics"},"type":{"example":"personal","allOf":[{"$ref":"#/components/schemas/ServerType"}]},"icon":{"type":"string","format":"uri","example":"https://cdn.example.com/icons/cs.png","description":"PUBLIC HTTPS(S) URL to an icon"},"degreeId":{"type":"string","description":"Associated degree identifier"},"degreeModuleId":{"type":"string","description":"Associated degree-module identifier"}},"required":["name","type","degreeId","degreeModuleId"]},"SocietyCategory":{"type":"string","enum":["Sports & Fitness","Academic & Professional","Arts & Culture","Community & Lifestyle"],"description":"Category label — used to group citysocieties servers"},"ServerViewDto":{"type":"object","properties":{"id":{"type":"string"},"name":{"type":"string"},"type":{"allOf":[{"$ref":"#/components/schemas/ServerType"}]},"slug":{"type":"string"},"degreeId":{"type":"string"},"degreeModuleId":{"type":"string"},"moduleYear":{"type":"number","description":"Academic year of the degree module (1-based)"},"ownerId":{"type":"string","description":"BetterAuth user id when the server has an owner"},"icon":{"type":"string"},"category":{"description":"Category label — used to group citysocieties servers","allOf":[{"$ref":"#/components/schemas/SocietyCategory"}]},"createdAt":{"type":"string"},"updatedAt":{"type":"string"}},"required":["id","name","type","slug","degreeId","degreeModuleId","createdAt","updatedAt"]},"ServerListResponseDto":{"type":"object","properties":{"items":{"description":"list of servers","type":"array","items":{"$ref":"#/components/schemas/ServerViewDto"}},"total":{"type":"number","example":42},"page":{"type":"number","example":1},"pageSize":{"type":"number","example":20}},"required":["items","total"]},"UpdateServerDto":{"type":"object","properties":{"name":{"type":"string","minLength":2,"maxLength":60,"example":"Calisthenics","description":"We discuss calisthenics"},"type":{"example":"personal","allOf":[{"$ref":"#/components/schemas/ServerType"}]},"icon":{"type":"string","format":"uri","example":"https://cdn.example.com/icons/cs.png","description":"PUBLIC HTTPS(S) URL to an icon"},"degreeId":{"type":"string","description":"Associated degree identifier"},"degreeModuleId":{"type":"string","description":"Associated degree-module identifier"}}},"AddOwnerDto":{"type":"object","properties":{"email":{"type":"string","format":"email","example":"student@city.ac.uk"}},"required":["email"]},"CreateNotificationDto":{"type":"object","properties":{"userId":{"type":"string","description":"Recipient BetterAuth user id","example":"usr_01hxt8zshm8yc6a5n8s6k1qj3r"},"channelId":{"type":"string","description":"Channel identifier (if applicable)"},"serverId":{"type":"string","description":"Server identifier (if applicable)"},"serverName":{"type":"string","description":"Server name (if applicable)"},"type":{"type":"string","description":"Notification type","example":"message.mention","enum":["message.create","message.mention","channel.invite","membership.status","generic"]},"title":{"type":"string","description":"Primary title for display","example":"New mention in #freshers"},"body":{"type":"string","description":"Optional body or preview text","example":"Alice mentioned you in #freshers"},"actorId":{"type":"string","description":"Actor user id triggering the notification","example":"usr_01hxt8zshm8yc6a5n8s6k1qabc"},"data":{"type":"object","description":"Structured context payload (server/channel/message ids)"},"expiresAt":{"type":"string","description":"Expiry timestamp; notification auto-deletes afterwards","example":"2024-05-20T00:00:00.000Z"},"dedupeKey":{"type":"string","description":"Globally-unique dedupe key; duplicate creates with the same key are silently dropped"}},"required":["userId","type","title"]},"NotificationViewDto":{"type":"object","properties":{"id":{"type":"string","description":"Notification identifier","example":"ntf_01hyf4a12x3n8zg9c0p7"},"userId":{"type":"string","description":"Recipient user id","example":"usr_01hxt8zshm8yc6a5n8s6k1qj3r"},"actorId":{"type":"string","description":"Actor user id if applicable","example":"usr_01hxt8zshm8yc6a5n8s6k1qabc"},"serverId":{"type":"string","description":"Server identifier associated with this notification","example":"srv_01hxt8zshm8yc6a5n8s6k1qz9x"},"serverName":{"type":"string","description":"Server name for display","example":"Physics 101"},"channelId":{"type":"string","description":"Channel identifier if the notification targets a channel","example":"chn_01hxt8zshm8yc6a5n8s6k1qy1m"},"type":{"type":"string","description":"Notification type discriminator","example":"message.mention","enum":["message.create","message.mention","channel.invite","membership.status","generic"]},"title":{"type":"string","description":"Display title","example":"New mention in #freshers"},"body":{"type":"string","description":"Optional body/preview","example":"Alice mentioned you in #freshers"},"status":{"type":"string","description":"Current status","enum":["unread","read"],"example":"unread"},"data":{"type":"object"},"createdAt":{"type":"string","description":"Creation timestamp","example":"2024-05-16T09:15:30.000Z"},"updatedAt":{"type":"string","description":"Last update timestamp","example":"2024-05-16T09:15:30.000Z"},"readAt":{"type":"string","description":"Read timestamp if set","example":"2024-05-16T10:00:00.000Z"},"seenAt":{"type":"string","description":"Seen timestamp for badge counts","example":"2024-05-16T10:05:00.000Z"},"expiresAt":{"type":"string","description":"Expiry timestamp (auto-clean via TTL)","example":"2024-05-20T00:00:00.000Z"}},"required":["id","userId","type","title","status","createdAt","updatedAt"]},"NotificationPreferenceViewDto":{"type":"object","properties":{"channelId":{"type":"string","description":"Target channel id"},"level":{"type":"string","description":"Notification level for this channel","enum":["all","mentions","none"]}},"required":["channelId","level"]},"UpdateNotificationPreferenceDto":{"type":"object","properties":{"level":{"type":"string","description":"Notification level for this channel","enum":["all","mentions","none"],"example":"all"}},"required":["level"]},"CreateChannelDto":{"type":"object","properties":{"name":{"type":"string","minLength":1,"maxLength":32},"type":{"type":"string","enum":["text","qa"]},"privacy":{"type":"string","enum":["public","hidden"]},"position":{"type":"number","minimum":0},"memberIds":{"description":"Optional seed access list for *hidden* channels.\nBA user IDs (strings), not ObjectIds.","uniqueItems":true,"type":"array","items":{"type":"string"}}},"required":["name","type","privacy"]},"ChannelViewDto":{"type":"object","properties":{"id":{"type":"string","description":"Channel id","example":"6710a4f2f23a5a2c5f6f8a1b"},"serverId":{"type":"string","description":"Server id this channel belongs to","example":"670ffe9bd4e2f1d3c9a4b123"},"name":{"type":"string","description":"Channel name (unique within server, unless you allow dupes)","example":"general","minLength":1,"maxLength":50},"type":{"type":"string","description":"Channel type","enum":["text","qa"],"example":"text"},"position":{"type":"number","description":"Ordering position within the server","example":0},"privacy":{"type":"string","description":"Channel privacy","enum":["public","hidden"],"example":"public"},"createdAt":{"type":"string","description":"Creation timestamp (ISO 8601)","example":"2025-09-27T15:21:45.123Z","format":"date-time"},"updatedAt":{"type":"string","description":"Last update timestamp (ISO 8601)","example":"2025-10-01T10:05:12.987Z","format":"date-time"}},"required":["id","serverId","name","type","position","privacy","createdAt","updatedAt"]},"ChannelListResponseDto":{"type":"object","properties":{"publicChannels":{"description":"the public channels for the server","items":{"type":"array"},"type":"array"},"privateChannels":{"description":"private channels the user has joined","type":"array","items":{"$ref":"#/components/schemas/ChannelViewDto"}},"total":{"type":"number","example":42},"page":{"type":"number","example":1},"pageSize":{"type":"number","example":20}},"required":["publicChannels","total","page","pageSize"]},"CreateAttachmentDto":{"type":"object","properties":{"url":{"type":"string","description":"Public URL pointing to stored attachment","example":"https://cdn.example.com/uploads/file.pdf"},"name":{"type":"string","description":"Original filename if provided","example":"lecture-notes.pdf"},"mime":{"type":"string","description":"Mime type of the attachment","example":"application/pdf"},"size":{"type":"number","description":"File size in bytes","example":1048576}},"required":["url"]},"CreateMentionDto":{"type":"object","properties":{"userId":{"type":"string","description":"Identifier of the mentioned user (BetterAuth user id)","example":"usr_01hxt8zshm8yc6a5n8s6k1qj3r"}},"required":["userId"]},"CreateMessageDto":{"type":"object","properties":{"channelId":{"type":"string","description":"Target channel identifier","example":"6710a4f2f23a5a2c5f6f8a1b"},"content":{"type":"string","maxLength":4000,"description":"Message Markdown/text content","example":"Remember to review the lecture slides before Friday's quiz."},"authorName":{"type":"string","maxLength":200,"description":"Name of the author (defaults to session user name)","example":"Aaliyah Khan"},"attachments":{"description":"List of attachments associated with this message","type":"array","items":{"$ref":"#/components/schemas/CreateAttachmentDto"}},"mentions":{"description":"Users mentioned in the message content","type":"array","items":{"$ref":"#/components/schemas/CreateMentionDto"}}},"required":["channelId","content"]},"MessageViewDto":{"type":"object","properties":{"id":{"type":"string","description":"Message identifier","example":"6750a4f2f23a5a2c5f6f8a1b"},"channelId":{"type":"string","description":"Channel identifier this message belongs to","example":"6710a4f2f23a5a2c5f6f8a1b"},"authorId":{"type":"string","description":"Author identifier (BetterAuth user id)","example":"usr_01hxt8zshm8yc6a5n8s6k1qj3r"},"authorName":{"type":"string","description":"Name of the Author","example":"Khan"},"content":{"type":"string","description":"Message Markdown/text content","maxLength":4000},"attachments":{"description":"Attachments included with the message","items":{"type":"array"},"type":"array"},"mentions":{"description":"User mentions referenced in the message","items":{"type":"array"},"type":"array"},"editedAt":{"type":"string","description":"Timestamp of last edit in ISO 8601 format","example":"2025-10-01T10:05:12.987Z","format":"date-time"},"createdAt":{"type":"string","description":"Message creation timestamp (ISO 8601)","example":"2025-09-27T15:21:45.123Z","format":"date-time"},"updatedAt":{"type":"string","description":"Last update timestamp (ISO 8601)","example":"2025-10-01T11:05:12.987Z","format":"date-time"}},"required":["id","channelId","authorId","authorName","content","createdAt","updatedAt"]},"MessageListResponseDto":{"type":"object","properties":{"items":{"description":"Messages ordered chronologically (oldest to newest unless specified)","items":{"type":"array"},"type":"array"},"total":{"type":"number","description":"Total messages available for the current filter","example":125},"page":{"type":"number","description":"Current page when using page-based pagination","example":1},"pageSize":{"type":"number","description":"Number of messages per page","example":50},"hasMore":{"type":"boolean","description":"Whether additional pages are available","example":false}},"required":["items","total","page","pageSize","hasMore"]},"DegreeViewDto":{"type":"object","properties":{"id":{"type":"string","description":"Degree identifier"},"slug":{"type":"string","description":"URL-friendly identifier"},"name":{"type":"string","description":"Degree name"},"durationYears":{"type":"number","description":"duration of the years"},"type":{"type":"string","description":"type of degree","enum":["undergraduate"]}},"required":["id","slug","name","durationYears","type"]},"DegreeModuleViewDto":{"type":"object","properties":{"id":{"type":"string"},"moduleId":{"type":"string"},"title":{"type":"string"},"kind":{"type":"string","enum":["core","elective"]},"term":{"type":"string","enum":["firstterm","secondterm","full-year"]},"year":{"type":"number"},"description":{"type":"string"},"credits":{"type":"number"}},"required":["id","moduleId","title","kind","year"]},"DegreeDetailDto":{"type":"object","properties":{"id":{"type":"string","description":"Degree identifier"},"slug":{"type":"string","description":"URL-friendly identifier"},"name":{"type":"string","description":"Degree name"},"durationYears":{"type":"number","description":"duration of the years"},"type":{"type":"string","description":"type of degree","enum":["undergraduate"]},"modules":{"type":"array","items":{"$ref":"#/components/schemas/DegreeModuleViewDto"}}},"required":["id","slug","name","durationYears","type","modules"]}}}}